Support retrieval of search statistic data via web api
