+++ app/frontend/components/verse_statistics_chart.js
@@ -117,7 +117,7 @@ class VerseStatisticsChart {
     var currentTranslation = app_controller.tab_controller.getTab(tabIndex).getBibleTranslationId();
     var bookList = await ipcNsi.getBookList(currentTranslation);
 
-    const [labels, values] = await ipcGeneral.getSearchStatisticChartData(currentTranslation, bookList, bibleBookStats);
+    const [labels, values] = await ipcNsi.getSearchStatisticChartData(currentTranslation, bookList, bibleBookStats);
 
     var data = {
       labels: labels,
